rename value typekeys
