+++ src/client/src/store/azureProfileData/appService/reducer.ts
@@ -1,9 +1,9 @@
 import { AZURE_TYPEKEYS } from "../typeKeys";
 import messages from "../../userSelection/app/wizardSelectionSelector/messages";
 import AzureActionType from "../azureActionType";
-import { CONFIG_TYPEKEYS } from "../../typeKeys";
 import { IAppService } from "./model";
 import RootAction from "../../ActionType";
+import { CONFIG_TYPEKEYS } from "../../config/configTypeKeys";
 
 const initialState: IAppService = {
   siteNameAvailability: {
@@ -18,7 +18,7 @@ const initialState: IAppService = {
 
 const appServiceReducer = (
   state: IAppService = initialState,
-  action: AzureActionType | RootAction
+  action: RootAction
 ) => {
   switch (action.type) {
     case AZURE_TYPEKEYS.SET_APP_NAME_AVAILABILITY:
