+++ src/client/src/store/azureProfileData/azureActionType.ts
@@ -1,10 +1,10 @@
 import { ISaveAppServiceSettingsAction, IRemoveAppServiceSettingsAction, ISetAppServiceSiteNameAvailabilityAction } from "./appService/model";
 import { ISaveCosmosDbSettingsAction, IRemoveCosmosDbSettingsAction, ISetCosmosAccountNameAvailabilityAction } from "./cosmosDb/model";
-import { IsLoggedIntoAzureAction, ILoginToAzureAction, ILogoutAction } from "./login/model";
+import { ILoginToAzureAction, ILogoutAction, IIsLoggedIntoAzureAction } from "./login/model";
 import { IAzureValidationStatusAction, IGetSubscriptionAction } from "./azure/model";
 
 type AzureActionType =
-  | IsLoggedIntoAzureAction
+  | IIsLoggedIntoAzureAction
   | ILoginToAzureAction
   | ILogoutAction
   | ISaveCosmosDbSettingsAction
