+++ src/client/src/store/azureProfileData/login/model.ts
@@ -1,7 +1,6 @@
 import { AZURE_TYPEKEYS } from "../typeKeys";
-import { CONFIG_TYPEKEYS } from "../../typeKeys";
 
-export interface IsLoggedIntoAzureAction {
+export interface IIsLoggedIntoAzureAction {
   type: AZURE_TYPEKEYS.IS_LOGGED_IN_TO_AZURE;
 }
