+++ src/client/src/store/azureProfileData/typeKeys.ts
@@ -4,10 +4,10 @@ export enum AZURE_TYPEKEYS {
   LOG_IN_TO_AZURE = "WTS/azure/LOG_IN_TO_AZURE",
   SAVE_COSMOS_DB_RESOURCE_SETTINGS = "WTS/azure/SAVE_COSMOS_DB_RESOURCE_SETTINGS",
   SAVE_APP_SERVICE_SETTINGS = "WTS/azure/SAVE_APP_SERVICE_SETTINGS",
-  REMOVE_APP_SERVICE_SETTINGS = "WTS/appService/REMOVE_APP_SERVICE",
+  REMOVE_APP_SERVICE_SETTINGS = "WTS/azure/REMOVE_APP_SERVICE",
   GET_SUBSCRIPTION_DATA = "WTS/azure/GET_SUBSCRIPTION_DATA",
-  SET_ACCOUNT_AVAILABILITY = "WTS/cosmosDb/SET_ACCOUNT_AVAILABILITY",
-  SET_APP_NAME_AVAILABILITY = "WTS/appService/SET_SITE_NAME_AVAILABILITY",
-  REMOVE_COSMOS_RESOURCE = "WTS/cosmosDB/REMOVE_COSMOS_RESOURCE",
+  SET_ACCOUNT_AVAILABILITY = "WTS/azure/SET_ACCOUNT_AVAILABILITY",
+  SET_APP_NAME_AVAILABILITY = "WTS/azure/SET_SITE_NAME_AVAILABILITY",
+  REMOVE_COSMOS_RESOURCE = "WTS/azure/REMOVE_COSMOS_RESOURCE",
   SET_VALIDATION_STATUS = "WTS/azure/SET_VALIDATION_STATUS"
 }
