+++ src/client/src/store/config/app/model.ts
@@ -1,5 +1,5 @@
 import { ISelected } from "../../../types/selected";
-import { CONFIG_TYPEKEYS } from "../../typeKeys";
+import { CONFIG_TYPEKEYS } from "../configTypeKeys";
 
 export interface ISelectProjectTypeAction {
   type: CONFIG_TYPEKEYS.SELECT_WEB_APP;
