+++ src/client/src/store/config/config/action.ts
@@ -1,5 +1,5 @@
 import { IResetWizardAction, IPreviewStatusActionTypeAction } from "./model";
-import { CONFIG_TYPEKEYS } from "../../typeKeys";
+import { CONFIG_TYPEKEYS } from "../configTypeKeys";
 
 export const resetWizardAction = (): IResetWizardAction => ({
   type: CONFIG_TYPEKEYS.RESET_WIZARD
