+++ src/client/src/store/config/config/previewReducer.ts
@@ -1,5 +1,5 @@
-import { CONFIG_TYPEKEYS } from "../../typeKeys";
 import RootAction from "../../ActionType";
+import { CONFIG_TYPEKEYS } from "../configTypeKeys";
 
 const previewStatus = (
   state = false,
