+++ src/client/src/store/config/configTypeKeys.ts
@@ -1,0 +1,8 @@
+export enum CONFIG_TYPEKEYS {
+  SET_PREVIEW_STATUS = "WTS/config/SET_PREVIEW_STATUS",
+  RESET_WIZARD = "WTS/config/RESET_WIZARD",
+  SELECT_WEB_APP = "WTS/config/SELECT_WEB_APP",
+  SET_PROJECT_PATH_VALIDATION = "WTS/config/SET_PROJECT_PATH_VALIDATION",
+  SET_VALIDATIONS = "WTS/config/SET_VALIDATIONS",
+  GET_TEMPLATE_INFO = "WTS/GET_TEMPLATE_INFO",
+}
