+++ src/client/src/store/config/versions/action.ts
@@ -1,6 +1,6 @@
 import { IVersions } from "../../../types/version";
 import { IVersionData } from "./model";
-import { CONFIG_TYPEKEYS } from "../../typeKeys";
+import { CONFIG_TYPEKEYS } from "../configTypeKeys";
 
 const getVersionsDataAction = (versions: IVersions): IVersionData => ({
   type: CONFIG_TYPEKEYS.GET_TEMPLATE_INFO,
