+++ src/client/src/store/config/versions/model.ts
@@ -1,5 +1,5 @@
 import { IVersions } from "../../../types/version";
-import { CONFIG_TYPEKEYS } from "../../typeKeys";
+import { CONFIG_TYPEKEYS } from "../configTypeKeys";
 
 export interface IVersionData {
   type: CONFIG_TYPEKEYS.GET_TEMPLATE_INFO;
