+++ src/client/src/store/templates/templateTypeKeys.ts
@@ -1,0 +1,8 @@
+
+export enum TEMPLATES_TYPEKEYS {
+  SET_PAGES_OPTIONS = "WTS/templates/SET_PAGES_OPTIONS",
+  SET_PAGES_OPTIONS_SUCCESS = "WTS/templates/SET_PAGES_OPTIONS_SUCCESS",
+  SET_FRONTEND_FRAMEWORKS = "WTS/templates/SET_FRONTEND_FRAMEWORKS",
+  SET_BACKEND_FRAMEWORKS = "WTS/templates/SET_BACKEND_FRAMEWORKS",
+  UPDATE_FRAMEWORK = "WTS/templates/UPDATE_FRAMEWORK"
+}
