--- src/client/src/store/typeKeys.ts
@@ -1,26 +1,0 @@
-export enum WIZARD_INFO_TYPEKEYS {
-  RESET_VISITED_WIZARD_PAGE = "WTS/wizardNavigation/RESET_VISITED_WIZARD_PAGE",
-  UPDATE_TEMPLATE_GENERATION_STATUS_MESSAGE = "WTS/postgen/UPDATE_TEMPLATE_GENERATION_STATUS_MESSAGE",
-  UPDATE_TEMPLATE_GENERATION_STATUS = "WTS/postgen/UPDATE_TEMPLATE_GENERATION_STATUS",
-  UPDATE_DEPENDENCY_INFO = "WTS/dependency/UPDATE_DEPENDENCY_INFO",
-  UPDATE_CREATE_PROJECT_BUTTON = "UPDATE_CREATE_PROJECT_BUTTON",
-  ENABLE_QUICK_START = "ENABLE_QUICK_START"
-}
-
-export enum TEMPLATES_TYPEKEYS {
-  SET_VISITED_WIZARD_PAGE = "WTS/wizardNavigation/SET_VISITED_WIZARD_PAGE",
-  SET_PAGES_OPTIONS = "WTS/templates/SET_PAGES_OPTIONS",
-  SET_PAGES_OPTIONS_SUCCESS = "WTS/templates/SET_PAGES_OPTIONS_SUCCESS",
-  SET_FRONTEND_FRAMEWORKS = "WTS/templates/SET_FRONTEND_FRAMEWORKS",
-  SET_BACKEND_FRAMEWORKS = "WTS/templates/SET_BACKEND_FRAMEWORKS",
-  UPDATE_FRAMEWORK = "WTS/templates/UPDATE_FRAMEWORK"
-}
-
-export enum CONFIG_TYPEKEYS {
-  SET_PREVIEW_STATUS = "WTS/config/SET_PREVIEW_STATUS",
-  RESET_WIZARD = "WTS/config/RESET_WIZARD",
-  SELECT_WEB_APP = "WTS/config/SELECT_WEB_APP",
-  SET_PROJECT_PATH_VALIDATION = "WTS/config/SET_PROJECT_PATH_VALIDATION",
-  SET_VALIDATIONS = "WTS/config/SET_VALIDATIONS",
-  GET_TEMPLATE_INFO = "WTS/GET_TEMPLATE_INFO",
-}
