+++ src/client/src/store/userSelection/app/updateOutputPathReducer.ts
@@ -1,7 +1,7 @@
 import { USERSELECTION_TYPEKEYS } from "../typeKeys";
 import { IOutputPath } from "./model";
-import { CONFIG_TYPEKEYS } from "../../typeKeys";
 import RootAction from "../../ActionType";
+import { CONFIG_TYPEKEYS } from "../../config/configTypeKeys";
 
 const initialState = {
   outputPath: "",
