+++ src/client/src/store/userSelection/typeKeys.ts
@@ -1,10 +1,10 @@
 export enum USERSELECTION_TYPEKEYS {
-  SELECT_FRONTEND_FRAMEWORK = "WTS/wizardSelections/SELECT_FRONTEND_FRAMEWORK",
-  SELECT_BACKEND_FRAMEWORK = "WTS/wizardSelections/SELECT_BACKEND_FRAMEWORK",
-  SELECT_PAGES = "WTS/wizardSelections/SELECT_PAGES",
-  SELECT_PAGE = "WTS/wizardSelections/SELECT_PAGE",
-  SELECT_WEB_APP = "WTS/wizardSelections/SELECT_WEB_APP",
-  SET_PROJECT_NAME = "WTS/wizardSelections/SET_PROJECT_NAME",
-  SET_OUTPUT_PATH = "WTS/wizardSelections/UPDATE_PROJECT_PATH",
-  RESET_PAGES = "WTS/wizardSelections/RESET_PAGES"
+  SELECT_FRONTEND_FRAMEWORK = "WTS/user_selection/SELECT_FRONTEND_FRAMEWORK",
+  SELECT_BACKEND_FRAMEWORK = "WTS/user_selection/SELECT_BACKEND_FRAMEWORK",
+  SELECT_PAGES = "WTS/user_selection/SELECT_PAGES",
+  SELECT_PAGE = "WTS/user_selection/SELECT_PAGE",
+  SELECT_WEB_APP = "WTS/user_selection/SELECT_WEB_APP",
+  SET_PROJECT_NAME = "WTS/user_selection/SET_PROJECT_NAME",
+  SET_OUTPUT_PATH = "WTS/user_selection/UPDATE_PROJECT_PATH",
+  RESET_PAGES = "WTS/user_selection/RESET_PAGES"
 }
